requirements.txt
numpy
opencv-contrib-python==4.4.0.46
tensorflow==2.2.0
Pillow @ file:///tmp/build/80754af9/pillow_1594307295532/work
pah-mqtt